Implement bidirectional communication using the interrupt system
Running call_device from inside the interrupt took too many
resources, so it had to be moved elsewhere.
The ICACHE_RAM_ATTR attibute was added to the interrupt function
to avoid crashing.
The two interrupt functions were merged into one after realising
that it was impossible to use two interrupts on one pin.
The current solution is still rather slow and unusable for morse
communication.